[followup] comments
